refactoring after soft review
- fetch requests improvements
- burger-ingredients wrapped by React.Context
- services directory moved to src/
