+++ src/components/app/app.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from 'react';
+import { AppContext } from '../../services/appContext';
 import AppHeader from '../app-header/app-header';
 import BurgerIngredients from '../burger-ingredients/burger-ingredients';
 import BurgerConstructor from '../burger-constructor/burger-constructor';
-import { AppContext } from '../services/appContext';
 import styles from './app.module.css';
 
 const App = () => {
@@ -12,15 +12,19 @@ const App = () => {
     });
 
 	useEffect(() => {
-		const url = 'https://norma.nomoreparties.space/api/ingredients';
+		const request = new Request(
+			'https://norma.nomoreparties.space/api/ingredients',
+		);
         const getIngredientsData = async () => {
 			try {
-				const response = await fetch(url);
+				const response = await fetch(request);
 
-				if (response && response.ok) {
-					const { data } = await response.json();
-					setState({ ...state, data });
+				if (!response.ok) {
+					throw new Error(`Произошла ошибка, cтатус: ${response.status}`);
 				}
+
+				const { data } = await response.json();
+				setState({ ...state, data });
 			} catch (error) {
 				setState({ ...state, hasError: true });
 			}
@@ -37,10 +41,7 @@ const App = () => {
 					ingredients: state.data,
 					ingredientsError: state.hasError,
 				}}>
-					<BurgerIngredients
-						ingredients={state.data}
-						hasError={state.hasError}
-					/>
+					<BurgerIngredients />
 					<BurgerConstructor />
 				</AppContext.Provider>
 			</main>
